+++ apps/files/appinfo/application.php
@@ -1,6 +1,5 @@
 <?php
 /**
- * @author Morris Jobke <hey@morrisjobke.de>
  * @author Roeland Jago Douma <roeland@famdouma.nl>
  * @author Tobias Kaminsky <tobias@kaminsky.me>
  * @author Vincent Petry <pvince81@owncloud.com>
@@ -21,13 +20,13 @@
  * along with this program.  If not, see <http://www.gnu.org/licenses/>
  *
  */
-
 namespace OCA\Files\Appinfo;
 
 use OCA\Files\Controller\ApiController;
 use OCP\AppFramework\App;
 use \OCA\Files\Service\TagService;
 use \OCP\IContainer;
+use OCA\Files\Capabilities;
 
 class Application extends App {
 	public function __construct(array $urlParams=array()) {
@@ -67,6 +66,13 @@ class Application extends App {
 				$c->query('Tagger'),
 				$homeFolder
 			);
+		});
+
+		/*
+		 * Register capabilities
+		 */
+		$server->getCapabilitiesManager()->registerCapability(function() {
+			return new Capabilities();
 		});
 	}
 }
