+++ apps/files_sharing/appinfo/application.php
@@ -2,6 +2,7 @@
 /**
  * @author Morris Jobke <hey@morrisjobke.de>
  * @author Robin Appelman <icewind@owncloud.com>
+ * @author Roeland Jago Douma <roeland@famdouma.nl>
  *
  * @copyright Copyright (c) 2015, ownCloud, Inc.
  * @license AGPL-3.0
@@ -31,6 +32,7 @@ use OCA\Files_Sharing\Controllers\ExternalSharesController;
 use OCA\Files_Sharing\Controllers\ShareController;
 use OCA\Files_Sharing\Middleware\SharingCheckMiddleware;
 use \OCP\IContainer;
+use OCA\Files_Sharing\Capabilities;
 
 class Application extends App {
 	public function __construct(array $urlParams = array()) {
@@ -119,6 +121,15 @@ class Application extends App {
 			$server = $c->query('ServerContainer');
 			return new PropagationManager(
 				$server->getUserSession(),
+				$server->getConfig()
+			);
+		});
+
+		/*
+		 * Register capabilities
+		 */
+		$server->getCapabilitiesManager()->registerCapability(function() use ($server) {
+			return new Capabilities(
 				$server->getConfig()
 			);
 		});
